feat: updated ux
